feat: indices search algorithm

+++ internal/repos/indices.go
@@ -4,5 +4,5 @@ import "context"
 
 // Indices repository is responsible for searching indices of given int val
 type Indices interface {
-	Search(ctx context.Context, val int) (int, error)
+	Search(ctx context.Context, val int64) (int, error)
 }
